:truck: chore: 게임 로직 일부 수정
- CRUD Repository에서 지원하지 않는 메서드 수정
- RoleManager 로그 추가
- 익셉션 추가
- 더티체킹이 불가능한 Redis Hash 객체 저장 기능 추가
- 방장은 자동으로 플레이어 생성 및 추가

+++ src/main/java/com/springles/game/GameSessionManager.java
@@ -29,13 +29,16 @@ public class GameSessionManager {
 
     /* 게임 세션 생성 */
     public void createGame(ChatRoom chatRoom) {
-        gameSessionRedisRepository.save(GameSession.of(chatRoom));
+        GameSession gameSession = gameSessionRedisRepository.save(GameSession.of(chatRoom));
+        addUser(chatRoom.getId(), chatRoom.getOwnerId());
     }
 
     /* 게임 시작 */
     public void startGame(Long roomId) {
         List<Player> players = findPlayersByRoomId(roomId);
-        if (players.size()< 5 || players.size() > 10) throw new CustomException(ErrorCode.PLAYER_HEAD_ERROR);
+        if (players.size() < 5 || players.size() > 10) {
+            throw new CustomException(ErrorCode.PLAYER_HEAD_ERROR);
+        }
         roleManager.assignRole(players);
         GameSession gameSession = findGameByRoomId(roomId);
         gameSessionRedisRepository.save(gameSession.start(players.size()));
@@ -48,13 +51,19 @@ public class GameSessionManager {
         gameSession.end();
         gameSessionRedisRepository.save(gameSession);
         List<Player> players = findPlayersByRoomId(roomId);
-        for (Player player : players) player.updateRole(GameRole.NONE);
+        for (Player player : players) {
+            player.updateRole(GameRole.NONE);
+        }
+        playerRedisRepository.saveAll(players);
+
     }
 
     /* 게임 세션 삭제 */
     public void removeGame(Long roomId) {
         List<Player> players = findPlayersByRoomId(roomId);
-        playerRedisRepository.deleteAll(players);
+        if (!players.isEmpty()) {
+            throw new CustomException(ErrorCode.GAME_PLAYER_EXISTS);
+        }
         gameSessionRedisRepository.deleteById(roomId);
     }
 
@@ -64,9 +73,12 @@ public class GameSessionManager {
         GameSession gameSession = findGameByRoomId(roomId);
         playerRedisRepository.deleteById(playerId);
         List<Player> players = findPlayersByRoomId(roomId);
-        if (players.isEmpty()) removeGame(roomId); // 아무도 없다면 방삭제
+        // 아무도 없다면 방삭제
+        if (players.isEmpty()) {
+            removeGame(roomId);
+        }
+        // 남은 플레이어가 존재하고 방장이 나갔다면 랜덤으로 방장 넘겨주기
         else if (Objects.equals(gameSession.getHostId(), playerId)) {
-            // 랜덤으로 방장 넘겨주기
             Random random = new Random();
             gameSession.changeHost(players.get(random.nextInt(players.size())).getMemberId());
             gameSessionRedisRepository.save(gameSession);
@@ -75,14 +87,15 @@ public class GameSessionManager {
 
     /* 게임에 유저 추가 */
     public void addUser(Long roomId, Long memberId) {
-        if (playerRedisRepository.countByRoomId(roomId) > 10) throw new CustomException(ErrorCode.GAME_HEAD_FULL);
+        if (playerRedisRepository.findByRoomId(roomId).size() > 10) {
+            throw new CustomException(ErrorCode.GAME_HEAD_FULL);
+        }
         GameSession gameSession = findGameByRoomId(roomId);
         // 아직 다른 방에 참가중이라면 -> 게임 중간에 나갔을 경우 발생 가능 -> 중간에 나갔을 경우를 처리해야 함.
         if (playerRedisRepository.existsByMemberId(memberId)) {
-
+            throw new CustomException(ErrorCode.PLAYER_STILL_INGAME);
         }
-        Player newPlayer = Player.of(memberId, roomId);
-        playerRedisRepository.save(newPlayer);
+        playerRedisRepository.save(Player.of(memberId, roomId));
     }
 
     public GameSession findGameByRoomId(Long roomId) {
